Add "Save Current Positions"
Added Edit->Configure->Save Current Positions.  It saves the current Term and, if open, Edit window size and location information in the config file.

Tested on Windows 10, Ubuntu and Centos 7.

It has a minor annoyance on Windows in that I can't get the windows to move all the way to the left margin. There is a small gap.  The cause, as far as I can see, is that the xywh details provided by Qt for a maximized window return a negative xy. If I use negative values for xy in the config, the window is rendered mostly off screen.  The doc says negative values are relative to the right margin.  The problem does not seem to occur on Linux.